+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 0.7.14
+* Rewrote Expanded Arsenal to fix stacking bugs.
+
+# Version 0.7.13
+* Minor updates to spell lists to fix some missed cases.
+
 # Version 0.7.12
 * Fixed issue with missing log creators creating stutters and bloated logs.